Poc/kafka_anomaly_detection/requirements.txt
pandas
scikit-learn
mysql-connector-python
kafka-python
pymongo
numpy
tensorflow